Move sequence inspector view into module
